// planning-idsdp-pps-frontend/src/app/main/modules/feasibility-study-management/models/expenditure-cost.model.ts
export class ExpenditureCostModel {
    economicCodeId: number;
    economicSubCodeId: number;
    description: string;
    totalAmount: number;
    gobAmount: number;
    feGobAmount: number;
    paAmount: number;
    rpaAmount: number;
    rpaGobAmount: number;
    rpaSpecialAccountAmount: number;
    dpaAmount: number;
    dpaThroughPdAmount: number;
    dpaDpAmount: number;
    ownFundAmount: number;
    feOwnFundAmount: number;
    othersAmount: number;
    othersFeAmount: number;
    fspMasterId: number;
    uuid: any;
    id: number;


}